--- Project Phase 5/.gitignore
@@ -1,1 +1,0 @@
-/bin/
