+++ src/db/DBConnect.java
@@ -1,4 +1,4 @@
-package com.company;
+package db;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
