+++ web/WEB-INF/html/about.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <title>VGRentalz&trade;</title>
-	 <link rel="stylesheet" href="Styles/vgr.css">
+	 <link rel="stylesheet" href="../css/vgr.css">
       <meta charset="utf-8"> 
        <meta name ="description" content="Video Game rental website with 
                                        affordable pricing and premium perks
@@ -15,7 +15,7 @@
       
       <header>
           <h1 class="title"><a href="vgrFrontPage.html">VGRentalZ</a></h1> 
-	       <h2 class="page_header">GAMES</h2>
+          <h2 class="page_header">ABOUT US</h2>
           <ul class="navBar">
               <li class="navBar"><a href="registration.html">Registration</a></li>
               <li class="navBar"><a href="games.html">Games</a></li>
@@ -24,14 +24,23 @@
       </header>
       
       <main>	
-          <p class="category"><br><br><br>XBOX ONE</p>
-	       <img id ="vg_console" src="Images/xboxOne.jpg" alt="Xbox One" width="400" id="xboxOne" class="category">
+          <p class="description"><br><br><br>VGRentalZ is a website where anybody from the United States can rent physical copies of the hottest video games for only $2 a week. </p>
           
-          <p class="category"><br>PS4</p>
-	       <img id ="vg_console" src="Images/ps4.jpeg" alt="Xbox One" width="400" id="ps4" class="category">
+          <p class="features">Standard Features:</p>
+	       <ul>
+	           <li>Only $2 a week</li> 
+	           <li>No late fees! Pay weekly until you own the game or return it!</li>
+               <li>2 games per customer MAX at a time</li>
+               <li>Free shipping in 3-5 business days</li>
+	       </ul>
           
-          <p class="category"><br>NINTENDO SWITCH</p>
-	       <img id ="vg_console" src="Images/nintendoSwitch.webp" alt="Xbox One" width="400" id="switch" class="category">
+            <p class="features">Premium Features:</p>
+	           <ul>
+                   <li>Only $10 a month</li> 
+                   <li>$1.50/week instead of $2/week for standard members</li>
+                   <li>4 games per customer MAX at a time</li>
+                   <li>Free 2-day shipping guarantee!</li>
+	           </ul>
       </main>
       
       <footer>
