+++ web/WEB-INF/html/index.html
@@ -1,0 +1,132 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>VGRentalZ</title>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <link rel="stylesheet" href="../css/main.css">
+</head>
+<body>
+    <nav>
+        <div class="nav-wrapper primary1">
+            <a href="index.html" class="brand-logo">VGRentalZ</a>
+            <ul id="nav-mobile" class="right hide-on-med-and-down">
+                <li><a href="vgrFrontPage.html">Old</a></li>
+                <li><a href="vgrFrontPage.html">Game Library</a></li>
+                <li><a href="vgrFrontPage.html">Contact Us</a></li>
+                <li><a href="vgrFrontPage.html">Shopping Cart</a></li>
+            </ul>
+        </div>
+    </nav>
+
+<!-- Main container -->
+    <div class="container hide-on-small-only">
+        <p id="testMazie" class="z-depth-3">
+        </p>
+
+        <p>
+            <blockquote>
+                <h5>Our goal is to provide you with a wide variety of video games at an affordable price!</h5>
+            </blockquote>
+            <h6> Standard Features:</h6>
+            - Only $2 a week<br>
+            - No late fees! Pay weekly until you own the game or return it!<br>
+            - 2 games per customer MAX at a time<br>
+            - Free shipping in 3-5 business days<br><br>
+            PREMIUM FEATURES:<br>
+            -----------------------<br>
+            - Only $10 a month<br>
+            - $1.50/week instead of $2/week for standard members<br>
+            - 4 games per customer MAX at a time<br>
+            - Free 2-day shipping guarantee!
+        </p>
+
+        <!-- Basic card: about the company -->
+        <div class="row">
+            <div class="col s5">
+                <div class="card primary1">
+                    <div class="card-content white-text">
+                        <span class="card-title">Gaming with VGRentalZ</span>
+                        <p>
+                            Our goal is to provide you with a wide variety of video games at an affordable price!<br><br>
+                            STANDARD FEATURES:<br>
+                            --------------------------<br>
+                                - Only $2 a week<br>
+                                - No late fees! Pay weekly until you own the game or return it!<br>
+                                - 2 games per customer MAX at a time<br>
+                                - Free shipping in 3-5 business days<br><br>
+                            PREMIUM FEATURES:<br>
+                            -----------------------<br>
+                                - Only $10 a month<br>
+                                - $1.50/week instead of $2/week for standard members<br>
+                                - 4 games per customer MAX at a time<br>
+                                - Free 2-day shipping guarantee!
+                        </p>
+                    </div>
+                    <div class="card-action">
+                        <a href="#">This is a link</a>
+                        <a href="#">This is a link</a>
+                    </div>
+                </div>
+            </div>
+
+        <!-- Card: game library -->
+            <div class="col s6">
+                <div class="card">
+                    <div class="card-image">
+                        <img src="https://c1.wallpaperflare.com/preview/184/835/736/arcade-game-pacman-retro.jpg">
+                        <span class="card-title">Game Library</span>
+                    </div>
+                    <div class="card-content">
+                        <p> We carry games for the following gaming consoles:<br>
+                            - Xbox One<br>
+                            - PS4<br>
+                            - Nintendo Switch<br>
+                        </p>
+                    </div>
+                    <div class="card-action primary">
+                        <a href="#">Browse Game Library</a>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+    </div>
+
+<!--    <script-->
+<!--            src="https://code.jquery.com/jquery-3.4.1.slim.min.js"-->
+<!--            integrity="sha256-pasqAKBDmFT4eHoN2ndd6lN370kFiGUFyTiUHWhU7k8="-->
+<!--            crossorigin="anonymous"></script>-->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
+    <script src="../js/index.js"></script>
+</body>
+
+<!--<footer>-->
+<!--    <footer class="page-footer">-->
+<!--        <div class="container">-->
+<!--            <div class="row">-->
+<!--                <div class="col l6 s12">-->
+<!--                    <h5 class="white-text">Footer Content</h5>-->
+<!--                    <p class="grey-text text-lighten-4">You can use rows and columns here to organize your footer content.</p>-->
+<!--                </div>-->
+<!--                <div class="col l4 offset-l2 s12">-->
+<!--                    <h5 class="white-text">Links</h5>-->
+<!--                    <ul>-->
+<!--                        <li><a class="grey-text text-lighten-3" href="#!">Link 1</a></li>-->
+<!--                        <li><a class="grey-text text-lighten-3" href="#!">Link 2</a></li>-->
+<!--                        <li><a class="grey-text text-lighten-3" href="#!">Link 3</a></li>-->
+<!--                        <li><a class="grey-text text-lighten-3" href="#!">Link 4</a></li>-->
+<!--                    </ul>-->
+<!--                </div>-->
+<!--            </div>-->
+<!--        </div>-->
+<!--        <div class="footer-copyright">-->
+<!--            <div class="container">-->
+<!--                © 2014 Copyright Text-->
+<!--                <a class="grey-text text-lighten-4 right" href="#!">More Links</a>-->
+<!--            </div>-->
+<!--        </div>-->
+<!--    </footer>-->
+<!--</footer>-->
+</html>
