+++ web/WEB-INF/html/registration.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <title>VGRentalz&trade;</title>
-	 <link rel="stylesheet" href="Styles/vgr.css">
+	 <link rel="stylesheet" href="../css/vgr.css">
       <meta charset="utf-8"> 
        <meta name ="description" content="Video Game rental website with 
                                        affordable pricing and premium perks
