+++ web/WEB-INF/html/vgrFrontPage.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <title>VGRentalz&trade;</title>
-	 <link rel="stylesheet" href="Styles/vgr.css">
+	 <link rel="stylesheet" href="../css/vgr.css">
       <meta charset="utf-8"> 
        <meta name ="description" content="Video Game rental website with 
                                        affordable pricing and premium perks
@@ -17,6 +17,7 @@
           <h1 class="title"><a href="vgrFrontPage.html">VGRentalZ</a></h1> 
 	       <h2 class="page_header">Welcome to VGRentalZ where you can rent games for only $2 a week!</h2>
           <ul class="navBar">
+              <li class="navBar"><a href="index.html">MAZIE's HOME</a></li>
               <li class="navBar"><a href="registration.html">Registration</a></li>
               <li class="navBar"><a href="games.html">Games</a></li>
               <li class="navBar"><a href="about.html">About Us</a></li>
